adding frame collector and optional frame data from asset entity collector
