+++ openpype/settings/entities/schemas/projects_schema/schema_project_traypublisher.json
@@ -350,7 +350,7 @@
                     "name": "template_validate_plugin",
                     "template_data": [
                         {
-                            "key": "CollectFrameRangeData",
+                            "key": "CollectFrameDataFromAssetEntity",
                             "label": "Collect frame range from asset entity"
                         },
                         {
